requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
django-environ==0.4.5
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
line-bot-sdk==1.15.0
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
